+++ examples/rms.rs
@@ -30,14 +30,14 @@ async fn main() -> Result<()> {
 
         println!("Calculating rms around focus...");
         let mut sound_field = record
-            .rms(
+            .sound_field(
                 Range {
                     x: focus.x - 20.0..=focus.x + 20.0,
                     y: focus.y - 20.0..=focus.y + 20.0,
                     z: focus.z..=focus.z,
                     resolution: 1.,
                 },
-                RecordOption {
+                RmsRecordOption {
                     #[cfg(feature = "gpu")]
                     gpu: true,
                     ..Default::default()
@@ -75,14 +75,14 @@ async fn main() -> Result<()> {
 
         println!("Calculating rms with STM...");
         let mut sound_field = record
-            .rms(
+            .sound_field(
                 Range {
                     x: focus.x - 30.0..=focus.x + 30.0,
                     y: focus.y - 30.0..=focus.y + 30.0,
                     z: focus.z..=focus.z,
                     resolution: 1.,
                 },
-                RecordOption {
+                RmsRecordOption {
                     #[cfg(feature = "gpu")]
                     gpu: true,
                     ..Default::default()
