+++ src/lib.rs
@@ -6,9 +6,9 @@ mod utils;
 use autd3::controller::timer::TimerStrategy;
 use autd3::controller::ControllerBuilder;
 pub use error::EmulatorError;
-pub use option::{Range, RecordOption};
+pub use option::Range;
 use record::{DeviceRecord, TransducerRecord};
-pub use record::{Record, Rms, SoundField};
+pub use record::{Instant, InstantRecordOption, Record, Rms, RmsRecordOption};
 
 use std::time::Duration;
